代码页转换统一使用iconv，不再使用windows自带的 移除对BOOL,TRUE,FALSE的使用 将AddUnique，AddSorted，AddSortedUniqued，RemoveFromList，RemoveAllFromList，ExistInList，FindInList，IndexInList，ArrayMap，ReplaceInList这几个CEasyArray相关的全局函数改为了CEasyArray的成员函数AddUnique，AddSorted，AddSortedUniqued，Remove，RemoveAll，Exist，Find，FindIndex，Map，Replace 修复分组群发在多线程连接模式下无法送达的问题

+++ NetLib/ENLConnection.cpp
@@ -420,7 +420,7 @@ const BYTE* CENLConnection::DecompressMsg(const BYTE* pData, UINT& DataLen)
 				NULL);
 			if (Result == LZO_E_OK)
 			{
-				DataLen = OutLen;
+				DataLen = (UINT)OutLen;
 				return (BYTE*)m_CompressBuffer.GetBuffer();
 			}
 			else
